- When -k is used, also update the image date to match the date in iPhoto - Some fixes to correctly handle unicode file names (contributed by tom.insam) - If the image title already ends with .JPG, don't append another .jpg suffix on export   (contributed by tom.insam) - a few bug fixes here and there

+++ tilutil/exiftool.py
@@ -18,9 +18,11 @@ Created on May 29, 2009
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
+import datetime
 import os
 import sys
 import tempfile
+import time
 
 import xml.dom
 import xml.dom.minidom
@@ -54,11 +56,12 @@ copy from http://www.sno.phy.queensu.ca/~phil/exiftool/.
 def get_iptc_data(image_file):
   """get caption and keywords all in one operation."""
   output = systemutils.execandcombine(
-      "exiftool -X -m -q -q -Keywords -Caption-Abstract "
-      "'%s'" % (image_file))
+      "exiftool -X -m -q -q -Keywords -Caption-Abstract -DateTimeOriginal "
+      "'%s'" % (image_file.encode('utf8')))
 
   keywords = []
   caption = None
+  date_time_original = None
   if output:
     try:
       xml_data = xml.dom.minidom.parseString(output)
@@ -73,31 +76,50 @@ def get_iptc_data(image_file):
         for xml_caption in xml_data.getElementsByTagName(
             "IPTC:Caption-Abstract"):
           caption = xml_caption.firstChild.nodeValue
+        for xml_element in xml_data.getElementsByTagName(
+            "ExifIFD:DateTimeOriginal"):
+          date_time_original = time.strptime(xml_element.firstChild.nodeValue,
+                                             "%Y:%m:%d %H:%M:%S")
+          date_time_original = datetime.datetime(date_time_original.tm_year,
+                                                 date_time_original.tm_mon,
+                                                 date_time_original.tm_mday,
+                                                 date_time_original.tm_hour,
+                                                 date_time_original.tm_min,
+                                                 date_time_original.tm_sec)
     except xml.parsers.expat.ExpatError, ex:
       print >> sys.stderr, "Could not parse exiftool output %s: %s" % (
           output, ex)
 
-  return (keywords, caption)
+  return (keywords, caption, date_time_original)
 
 
-def update_iptcdata(filepath, new_caption, new_keywords): 
+def update_iptcdata(filepath, new_caption, new_keywords, new_datetime): 
   """Updates the caption and keywords of an image file."""
-  tmp = tempfile.mkstemp(dir="/var/tmp")[1]
-  file1 = open(tmp, "w")
-  if not new_caption:
-    # Setting the caption to an empty string for some reason dosn't work
-    new_caption = " "
-  print >> file1, new_caption.encode("utf-8")
-  file1.close()
   # Some cameras write into ImageDescription, so we wipe it out to not cause
   # conflicts with Caption-Abstract
-  command = (u'exiftool -F -m -P -ImageDescription= "-Caption-Abstract<=%s" ' %
-             (tmp))
-  for keyword in new_keywords:
-    command += u' -keywords="%s"' % (keyword)
+  command = u'exiftool -F -m -P -ImageDescription='
+  tmp = None
+  if not new_caption is None:
+    tmpfd, tmp = tempfile.mkstemp(dir="/var/tmp")
+    os.close(tmpfd)
+    file1 = open(tmp, "w")
+    if not new_caption:
+      # you can't set caption to an empty string
+      new_caption = " "
+    print >> file1, new_caption.encode("utf-8")
+    file1.close()
+    command += ' "-Caption-Abstract<=%s"' % (tmp)
+  
+  if new_datetime:
+    command += ' -DateTimeOriginal="%s"' % (
+      new_datetime.strftime("%Y:%m:%d %H:%M:%S"))
+  if new_keywords:
+    for keyword in new_keywords:
+      command += u' -keywords="%s"' % (keyword)
   command += ' "%s"' % (filepath)
   result = systemutils.execandcombine(command.encode("utf-8"))
-  os.remove(tmp)
+  if tmp:
+    os.remove(tmp)
   if result == "1 image files updated":
     # wipe out the back file created by exiftool
     backup_file = filepath + "_original"
